update low lv code

+++ LowLevel_Code/Master_Config_STM32H733/Core/Inc/AS5047U.h
@@ -65,7 +65,7 @@ typedef struct {
 	uint16_t CSGPIOPin;
 	uint16_t Position;
 	AS5047UError Error_Status;
-	float Offset;
+	double Offset;
 } AS5047U;
 
 
@@ -74,9 +74,9 @@ typedef struct {
 /*
  * Function
  */
-float EncPulse2Rad_Read(AS5047U *Enc,uint8_t inv_dir);
+double EncPulse2Rad_Read(AS5047U *Enc,uint8_t inv_dir);
 
-void AS5047U_init(AS5047U *dev, SPI_HandleTypeDef *hspiHandle, GPIO_TypeDef *CSGPIOTypedef, CRC_HandleTypeDef *hcrcHandle, uint16_t CSGPIOPin, float offset);
+void AS5047U_init(AS5047U *dev, SPI_HandleTypeDef *hspiHandle, GPIO_TypeDef *CSGPIOTypedef, CRC_HandleTypeDef *hcrcHandle, uint16_t CSGPIOPin, double offset);
 void AS5047U_Write(AS5047U *dev,uint16_t Register_Address, uint16_t Data);
 HAL_StatusTypeDef AS5047U_Read(AS5047U *dev,uint16_t Register_Address, uint16_t *Data);
 uint16_t AS5047U_Position_Highspeed_Read(AS5047U *dev,uint8_t dir);
