+++ LowLevel_Code/Master_Config_STM32H733/Core/Inc/kinematics.h
@@ -27,13 +27,13 @@ typedef enum
  * Function
  */
 
-Kinematics_StatusTypeDef IPK(float gammabar[3], float chi[4], float q[5]);
-Kinematics_StatusTypeDef IVK(float q[4], float chi_dot[3], float qv[4]);
-void FPK(float q[5], float l3, float Pne[3]);
-void FVK(float q[4], float qd[4], float l3, float twist[3]);
+Kinematics_StatusTypeDef IPK(double gammabar[3], double chi[4], double q[5]);
+Kinematics_StatusTypeDef IVK(double q[4], double chi_dot[3], double qv[4]);
+void FPK(double q[5], double l3, double Pne[3]);
+void FVK(double q[4], double qd[4], double l3, double twist[3]);
 
-void ChessPose(uint8_t Chess_Index, float Chess_Theta, float ChessPosition[2]);
-void FindR(float Offset2Center[2], float chessPosition[2], float *ChessRadius);
-void FindXYSpeed(float Chess_Theta, float Chess_Omaga, float ChessRadius, float SpeedXY[2]);
+void ChessPose(uint8_t Chess_Index, double Chess_Theta, double ChessPosition[2]);
+void FindR(double Offset2Center[2], double chessPosition[2], double *ChessRadius);
+void FindXYSpeed(double Chess_Theta, double Chess_Omaga, double ChessRadius, double SpeedXY[2]);
 
 #endif /* INC_KINEMATICS_H_ */
